+++ src/pages/artikel/detailarticle.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import DetailArticle from "@/components/page-artikel/DetailArticle";
+
+export default function detailarticle() {
+  return (
+    <div>
+      <DetailArticle />
+    </div>
+  );
+}
