+++ src/pages/artikel/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import HomeArticle from "@/components/HomeArticle";
+import HomeArticle from "@/components/page-artikel/HomeArticle";
 
 export default function index() {
   return (
